Code cleanup
- Rename Position to JobOpening
- Move to a separate module
